refactor(player-portal): decompose _quarantine/picker/ (phase 2)
picker/ pooled a generic UI primitive with character-domain
components. Splitting it:

- PickerDialog → shared/ui/ (modal shell with portal/backdrop/scroll
  lock; reusable across any feature that needs a centered modal).
- CompendiumPicker, CompendiumDetailPanel → characters/internal/
  (used by both creator and sheet — the only two pieces today that
  cross the creator/sheet seam).

The picker/index.ts barrel is gone; all consumers import the
component directly. _quarantine/picker/ is now empty.
