--- apps/player-portal/src/_quarantine/picker/index.ts
@@ -1,4 +1,0 @@
-export { CompendiumPicker, PickerResultList } from './CompendiumPicker';
-export type { CompendiumPickerProps, CompendiumPickerSplitPane, PickerResultListProps } from './CompendiumPicker';
-export { CompendiumDetailPanel } from './CompendiumDetailPanel';
-export { PickerDialog } from './PickerDialog';
