+++ apps/player-portal/src/features/characters/creator/AbilityBoostPicker.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import type { AbilityKey } from '@/features/characters/types';
 import { ABILITY_KEYS } from '@/features/characters/types';
 import { formatSignedInt } from '@/_quarantine/lib/format';
-import { PickerDialog } from '@/_quarantine/picker/PickerDialog';
+import { PickerDialog } from '@/shared/ui/PickerDialog';
 import type { CharacterContext } from '@/features/characters/internal/prereqs';
 
 interface Props {
