+++ apps/player-portal/src/features/characters/internal/CompendiumDetailPanel.tsx
